requirements.txt
streamlit
pandas
psycopg2-binary
nltk
scikit-learn
numpy
openai
python-dotenv
faiss-cpu
psutil
plotly